gpt-4o func and text2speech func
