--- gpt_copilot/test.json
@@ -1,34 +1,0 @@
-{
-    "path": "test.jpg",
-    "outputs": {
-        "object": [
-            {
-                "name": "roi",
-                "bndbox": {
-                    "xmin": 147,
-                    "ymin": 1482,
-                    "xmax": 769,
-                    "ymax": 1670
-                }
-            },
-            [
-                {
-                    "name": "warning",
-                    "bndbox": {
-                        "xmin": 159,
-                        "ymin": 1564,
-                        "xmax": 725,
-                        "ymax": 1662
-                    }
-                }
-            ]
-        ]
-    },
-    "time_labeled": 1722326502108,
-    "labeled": true,
-    "size": {
-        "width": 1734,
-        "height": 2312,
-        "depth": 3
-    }
-}
